requirements.txt
jinja2==3.1.4       # pyyaml dependency
lark==1.2.2         # pytest-cov dependency
pytest==8.3.2
pytest-cov==5.0.0
pyyaml==6.0.2
py-solc-x==2.0.3
typeguard==4.3.0    # pyyaml dependency
web3==7.5.0
